Changed type

+++ pointers_arrays_strings/8-print_diagsums.c
@@ -9,7 +9,7 @@
 
 void print_diagsums(int *a, int size)
 {
-long int i, sum1, sum2;
+int i, sum1, sum2;
 sum1 = 0;
 sum2 = 0;
 
@@ -18,5 +18,5 @@ for (i = 0; i <= (size * size); i = i + size + 1)
 for (i = 0; i <= (size * size) - size; i = i + size - 1)
 	sum2 = sum2 + *(a + i);
 
-printf("%li, %li\n", sum1, sum2);
+printf("%d, %d\n", sum1, sum2);
 }
